===== File Statistics: =====
Conflicts: 24
Deleted: 14
Added: 315
Updated: 929
==== Property Statistics: =====
Conflicts: 12
Updated: 34

==============================================================================================================================


CU /apps/drum/src/dspace-api/src/main/java/org/dspace/app/itemimport/ItemImport.java
C  /apps/drum/src/dspace-api/src/main/java/org/dspace/content/Community.java
C  /apps/drum/src/dspace-api/src/main/java/org/dspace/content/Bitstream.java
C  /apps/drum/src/dspace-api/src/main/java/org/dspace/content/Item.java
C  /apps/drum/src/dspace-api/src/main/resources/Messages.properties
C  /apps/drum/src/dspace-api/pom.xml
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/tools/edit-community.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/tools/edit-collection.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/submit/edit-metadata.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/submit/complete.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/layout/navbar-default.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/layout/navbar-admin.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/community-list.jsp
Skipped /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/help/collection-admin.html
Skipped /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/help/site-admin.html
C  /apps/drum/src/dspace-jspui/dspace-jspui-webapp/src/main/webapp/styles.css.jsp
C  /apps/drum/src/dspace-jspui/dspace-jspui-api/src/main/java/org/dspace/app/webui/servlet/admin/EditCommunitiesServlet.java
C  /apps/drum/src/dspace-jspui/dspace-jspui-api/src/main/java/org/dspace/app/webui/servlet/admin/AuthorizeAdminServlet.java
C  /apps/drum/src/dspace-jspui/dspace-jspui-api/src/main/java/org/dspace/app/webui/jsptag/ItemTag.java
C  /apps/drum/src/dspace-jspui/dspace-jspui-api/src/main/java/org/dspace/app/webui/util/Authenticate.java
C  /apps/drum/src/dspace-jspui/dspace-jspui-api/pom.xml
C  /apps/drum/src/pom.xml
C  /apps/drum/src/dspace/src/main/config/build.xml
C  /apps/drum/src/dspace/pom.xml
CC /apps/drum/src/dspace/bin/sub-daily
CC /apps/drum/src/dspace/bin/dsrun
MC /apps/drum/src/dspace/bin/create-administrator
MC /apps/drum/src/dspace/bin/index-update
MC /apps/drum/src/dspace/bin/update-handle-prefix
MC /apps/drum/src/dspace/bin/log-reporter
MC /apps/drum/src/dspace/bin/cleanup
MC /apps/drum/src/dspace/bin/start-handle-server
MC /apps/drum/src/dspace/bin/filter-media
MC /apps/drum/src/dspace/bin/dspace-info.pl
MC /apps/drum/src/dspace/bin/dsprop
MC /apps/drum/src/dspace/bin/make-handle-config

git-svn-id: https://svn.lib.umd.edu/drum/branches/drum-dspace-1.7.0/src@277 e8f66b91-402e-0410-ac90-b3834309c80d

+++ dspace-jspui/dspace-jspui-api/src/main/java/org/dspace/app/webui/util/FileUploadRequest.java
@@ -1,9 +1,9 @@
 /*
  * FileUploadRequest.java
  *
- * Version: $Revision: 3705 $
+ * Version: $Revision: 4243 $
  *
- * Date: $Date: 2009-04-11 19:02:24 +0200 (Sat, 11 Apr 2009) $
+ * Date: $Date: 2009-09-02 05:12:23 -0400 (Wed, 02 Sep 2009) $
  *
  * Copyright (c) 2002-2005, Hewlett-Packard Company and Massachusetts
  * Institute of Technology.  All rights reserved.
@@ -66,7 +66,7 @@ import org.dspace.core.ConfigurationManager;
  * allowable upload size.
  * 
  * @author Robert Tansley
- * @version $Revision: 3705 $
+ * @version $Revision: 4243 $
  */
 public class FileUploadRequest extends HttpServletRequestWrapper
 {
@@ -98,7 +98,7 @@ public class FileUploadRequest extends HttpServletRequestWrapper
         original = req;
 
         tempDir = ConfigurationManager.getProperty("upload.temp.dir");
-        int maxSize = ConfigurationManager.getIntProperty("upload.max");
+        long maxSize = ConfigurationManager.getLongProperty("upload.max");
 
         // Create a factory for disk-based file items
         DiskFileItemFactory factory = new DiskFileItemFactory();
